+++ README.md
@@ -5,4 +5,4 @@
 - [x] 레거시 컨트롤러의 반환 타입 원상복귀
 - [x] JspView 상수
 - [x] MappingRegistry가 Handler를 바로 꺼내주도록 수정
-- [ ] ControllerHandlerAdapter와 Registry 클래스의 패키지에 대한 고민
+- [x] ControllerHandlerAdapter와 Registry 클래스의 패키지에 대한 고민
