+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/HandlerAdapterRegistry.java
@@ -1,7 +1,6 @@
 package webmvc.org.springframework.web.servlet.mvc;
 
 import web.org.springframework.web.exception.InvalidHandlerTypeException;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerAdapter;
 
 import java.util.List;
